refactor: api 관련 로직 api 폴더로 이동
