+++ src/api/services/themes/index.ts
@@ -1,10 +1,10 @@
+import { BACKEND_API } from '@/api/instance';
 import { ERROR_MESSAGES } from '@/constants/errorMessage';
-import { BACKEND_API } from '@/services/api';
 import { ThemeCategoryData, ThemeHeroData } from '@/types/themeType';
 
 import { GetThemesResponse } from './types';
 
-export const fetchThemeData = async () => {
+export const fetchThemes = async () => {
   try {
     const response = await BACKEND_API.get<GetThemesResponse>('/api/v1/themes');
